cloud_functions/message_passing/ChatLogger/requirements.txt
google-cloud-firestore==2.9.0
Flask==2.2.5
functions-framework==3.0.0
